# src/docs/api.yaml
openapi: 3.0.0
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: ''
      tags:
        - Authentikáció
  /auth/login:
    post:
      operationId: login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '201':
          description: ''
      tags:
        - Authentikáció
...
info:
  title: RepVault
  description: |2-

          RepVault API
      
          A védett végpontok eléréséhez Bearer típusú JWT tokent kell használnia.
          Az authentikáció fül alatt lévő /login végponton keresztül tud tokenhez jutni. Adja meg a felhasználónevét és jelszavát,
          illetve ha még nincs felhasználója, regisztráljon az /register végponton keresztül és aztán jelentkezzen be. 
          Ezt követően kattintson az "Authorize" gombra a Swagger felületén, majd másolja be a tokenjét.
      
          Miután megadta a tokent, a Swagger automatikusan hozzáadja azt az "Authorization" fejléchez, 
          így Ön tesztelheti a védett végpontokat.
        
  version: '1.0'
  contact: {}
...
servers:
  - url: http://localhost:8000
    description: Development Server
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateUserDto:
      type: object
      properties: {}
    LoginDto:
      type: object
      properties: {}
...